// backup/features/account/types/types.ts
export interface Account {
    id: number
    created_at: string | Date
    updated_at: string | Date
    user_id: number
    type: string
    suscription_id: string | null
}

export interface Store {
    id: number
    name: string
    description: string | null
    logo: string | null
    slogan: string | null
    created_at: Date
    updated_at: Date
    user_id: number
    slug: string
    subdomain: string
}

export interface UserType {
    id: number
    email: string
    avatar: string | null
    banner: string | null
    created_at: string | Date
    first_name: string | null
    last_name: string | null
    username: string | null
    phone?: string | null
    password?: string
    updated_at?: string | Date
    supabase_user_id: string | null
    Account?: Account[]
    Store?: Store[]
}
// ...
export interface AccountDetailsTabProps {
    user: UserType
    translations: {
        "description.account-details": string
        "description.username": string
        "description.first-name": string
        "description.last-name": string
        "description.email": string
        "description.password": string
        "description.change-email": string
        "description.change-password": string
        "description.phone": string
        [key: string]: string
    }
    immediateData?: {
        username?: string | null
        email: string
        firstName?: string | null
        lastName?: string | null
        phone?: string | null
        avatar?: string | null
        createdAt: Date | string
        activeAccounts: number
        storesCount: number
        accountType: string
    }
    /*  onBannerUpdate?: (url: string | null) => void */
    onProfileUpdate: (profile: {
        username: string | null
        firstName: string | null
        lastName: string | null
        phone: string | null
    }) => void
}
// ...
export interface BasicInfoCardProps {
    user: UserType;
    immediateData?: AccountDetailsTabProps['immediateData'];
    onProfileUpdate: (profile: {
        username: string | null
        firstName: string | null
        lastName: string | null
        phone: string | null
    }) => void;
}